Fix temporary file is deleted before closing, in the rewrite function (#468)
Currently, if an error is raised while using files from the rewrite function, then the temporary file is deleted before closing it.

This is okay on unix, but unlinking open files causes an error on Windows.

+++ src/dotenv/main.py
@@ -1,6 +1,7 @@
 import io
 import logging
 import os
+import pathlib
 import shutil
 import sys
 import tempfile
@@ -131,17 +132,21 @@ def rewrite(
     path: StrPath,
     encoding: Optional[str],
 ) -> Iterator[Tuple[IO[str], IO[str]]]:
-    if not os.path.isfile(path):
-        with open(path, mode="w", encoding=encoding) as source:
-            source.write("")
+    pathlib.Path(path).touch()
+
     with tempfile.NamedTemporaryFile(mode="w", encoding=encoding, delete=False) as dest:
+        error = None
         try:
             with open(path, encoding=encoding) as source:
                 yield (source, dest)
-        except BaseException:
-            os.unlink(dest.name)
-            raise
-    shutil.move(dest.name, path)
+        except BaseException as err:
+            error = err
+
+    if error is None:
+        shutil.move(dest.name, path)
+    else:
+        os.unlink(dest.name)
+        raise error from None
 
 
 def set_key(
